pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "depth2room"
version = "0.1.0"
description = "Depth-conditioned video generation for room furnishing using VACE"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "einops",
    "accelerate",
    "numpy",
    "Pillow",
]

[project.optional-dependencies]
training = [
    "diffsynth-studio @ git+https://github.com/modelscope/DiffSynth-Studio.git",
    "wandb",
]
data = [
    "moderngl",
    "glcontext",
    "aiofiles",
    "aiohttp",
    "nest-asyncio",
    "transformers>=4.57.0",
    "kernels",
    "decord2",
]
inference = [
    "diffsynth-studio @ git+https://github.com/modelscope/DiffSynth-Studio.git",
    "imageio[ffmpeg]",
    "matplotlib",
    "scipy",
    "opencv-python",
    "OpenEXR",
    "safetensors",
    "av",
]
eval = [
    "lpips",
    "scikit-image",
    "opencv-python",
]
all = ["depth2room[training,data,inference,eval]"]

[tool.setuptools.packages.find]
where = ["src"]
